add: Decision maker accepts a querying agent
The querying is now a module itself. After each sequence the query method is updated with the posterior and returns the stimuli letters. Once the system commits to a decision (epoch), the agent updates itself and prepares for the new sequential decision making process.

Next step is to make disentangle the stimuli preparation from query selection.