Trigger handling (#185)
* Add new Trigger, TriggerHandler, TriggerType

New classes/methods for trigger handling

* Added some unit tests for new trigger

* Editing new Trigger unit tests

* Editing trigger and trigger tests

* Attempted fixing of some unit tests with varying success

* mock file open using unittest patch, uncomment old tests, add coverage

* lint, split up load method and define as staticmethod, updated variable names

* Edited triggers to catch some exceptions, added corresponding tests

* linting

* Trigger refinement based on feedback

Refactored type and TriggerType as category and TriggerCategory to avoid confusion with Python built-in type functions. Renamed FlushSensitivity to FlushFrequency. Added a docstring for TriggerHandler.

* Commented out dataclass import

So the linting tests will stop yelling at me

* Refining of Trigger based on feedback

Added dataclass back in to make Trigger frozen/(mostly) immutable, refactored Trigger.time as a float instead of a str

* Further refactoring of Trigger from suggestions

* linting

Co-authored-by: Tab Memmott <tab.memmott@gmail.com>